requirements.txt
asgiref==3.2.10
Django==3.0.8
future==0.18.2
python-whois==0.7.3
pytz==2020.1
sqlparse==0.3.1
